Rename Campaigns component to be called Campaign
This is a "step 2" in renaming campaign to Campaign. Git didn't see the
difference, so I changed the string in order to apply the capital "C".
